# swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/users/:userID": {
      "get": {
        "security": [
          {
            "api_key": [
              "[]"
            ]
          }
        ],
        "description": "\u003cb\u003eDescription\u003c/b\u003e\u003cbr\u003e\nReturns user by id\n\n\u003cbr\u003e\u003cb\u003e400: Bad Request\u003c/b\u003e\u003cbr\u003e\nErr400xBind_001 - Error while decoding request data to JSON - (could not decode parameters)\n\n\u003cbr\u003e\u003cb\u003e404: Not Found\u003c/b\u003e\u003cbr\u003e\nErr404xUser_001 - Error while getting user - (user not found)\n\n\u003cbr\u003e\u003cb\u003e500: Internal Server Error\u003c/b\u003e\u003cbr\u003e\nErr500xUser_001 - Error while getting user - (internal server error)\nErr500xRoles_001 - Error while getting user roles - (internal server error)\nErr500xTransaction_001 - Error while connecting to database - (server internal error)\nErr500xTransaction_002 - Error while committing all changes - (server internal error)",
        "tags": [
          "User"
        ],
        "summary": "Get user by id  ----- [Admin, MemberIsTarget]",
        "operationId": "userGet",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "userID",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "User data",
            "schema": {
              "$ref": "#/definitions/GetUserResponse"
            }
          },
          "default": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "api_key": [
              "[]"
            ]
          }
        ],
        "description": "\u003cb\u003eDescription\u003c/b\u003e\u003cbr\u003e\nDeletes user by ID\n\n\u003cbr\u003e\u003cb\u003e500: Internal Server Error\u003c/b\u003e\u003cbr\u003e\nErr500xUser_001 - Error deleting user - (internal server error)\nErr500xTransaction_001 - Error while connecting to database - (server internal error)\nErr500xTransaction_002 - Error while committing all changes - (server internal error)",
        "tags": [
          "User"
        ],
        "summary": "Delete user ----- [Admin, MemberIsTarget]",
        "operationId": "userDelete",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "User is deleted successfully",
            "schema": {
              "$ref": "#/definitions/DeleteUserResponse"
            }
          },
          "default": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "api_key": [
              "[]"
            ]
          }
        ],
        "description": "\u003cb\u003eDescription\u003c/b\u003e\u003cbr\u003e\nUpdates user by ID and returns updated user\n\n\u003cbr\u003e\u003cb\u003e400: Bad Request\u003c/b\u003e\u003cbr\u003e\nErr400x001 - Error while decoding request data to JSON - (could not decode parameters)\nErr400x002 - Error while verifying if request data has all required fields - (not all fields found)\nErr400x003 - User not authorized to edit password - (user not authorized)\n\n\u003cbr\u003e\u003cb\u003e404: Not Found\u003c/b\u003e\u003cbr\u003e\nErr404x001 - Error while getting user by email - (user not found)\nErr404x002 - Error while updating user - (user not found)\n\n\u003cbr\u003e\u003cb\u003e500: Internal Server Error\u003c/b\u003e\u003cbr\u003e\nErr500x001 - Error while getting user by email - (internal server error)\nErr500x002 - Error while hashing and setting new password - (internal server error)\nErr500x003 - Error while updating metadata - (internal server error)",
        "tags": [
          "User"
        ],
        "summary": "Update user  ----- [Admin, MemberIsTarget]",
        "operationId": "userUpdate",
        "parameters": [
          {
            "type": "string",
            "description": "User ID",
            "name": "userID",
            "in": "path"
          },
          {
            "description": "User model",
            "name": "request",
            "in": "body",
            "schema": {
              "description": "User model",
              "type": "object",
              "$ref": "#/definitions/UpdateUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User is updated successfully, user model",
            "schema": {
              "$ref": "#/definitions/UpdateUserResponse"
            }
          },
          "default": {
            "description": "genericError",
            "schema": {
              "$ref": "#/definitions/genericError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "DeleteUserResponse": {
      "type": "object",
      "properties": {
        "deleted": {
          "type": "boolean",
          "x-go-name": "Deleted"
        }
      },
      "x-go-package": "pr-reviewer/internal/api/v1/user"
    },
...
    "GetUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/User"
        }
      },
      "x-go-package": "pr-reviewer/internal/api/v1/user"
    },
...
    "Role": {
      "description": "Role is a function a user can serve",
      "type": "string",
      "x-go-package": "pr-reviewer/internal/model"
    },
...
    "UpdateUserRequest": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "displayName": {
          "type": "string",
          "x-go-name": "DisplayName"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "isCreator": {
          "type": "boolean",
          "x-go-name": "IsCreator"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "profileColorIndex": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ProfileColorIndex"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          },
          "x-go-name": "Roles"
        },
        "sequenceWallet": {
          "type": "string",
          "x-go-name": "SequenceWallet"
        },
        "stripeCustomerID": {
          "type": "string",
          "x-go-name": "StripeCustomerID"
        },
        "stripePaymentMethodID": {
          "type": "string",
          "x-go-name": "StripePaymentMethodID"
        },
        "userID": {
          "$ref": "#/definitions/UserID"
        }
      },
      "x-go-package": "pr-reviewer/internal/api/v1/user"
    },
    "UpdateUserResponse": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "displayName": {
          "type": "string",
          "x-go-name": "DisplayName"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "isCreator": {
          "type": "boolean",
          "x-go-name": "IsCreator"
        },
        "profileColorIndex": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ProfileColorIndex"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          },
          "x-go-name": "Roles"
        },
        "sequenceWallet": {
          "type": "string",
          "x-go-name": "SequenceWallet"
        },
        "stripeCustomerID": {
          "type": "string",
          "x-go-name": "StripeCustomerID"
        },
        "stripePaymentMethodID": {
          "type": "string",
          "x-go-name": "StripePaymentMethodID"
        },
        "userID": {
          "$ref": "#/definitions/UserID"
        }
      },
      "x-go-package": "pr-reviewer/internal/api/v1/user"
    },
    "User": {
      "description": "User is structure represent User object",
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "displayName": {
          "type": "string",
          "x-go-name": "DisplayName"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "isCreator": {
          "type": "boolean",
          "x-go-name": "IsCreator"
        },
        "profileColorIndex": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ProfileColorIndex"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Role"
          },
          "x-go-name": "Roles"
        },
        "sequenceWallet": {
          "type": "string",
          "x-go-name": "SequenceWallet"
        },
        "stripeCustomerID": {
          "type": "string",
          "x-go-name": "StripeCustomerID"
        },
        "stripePaymentMethodID": {
          "type": "string",
          "x-go-name": "StripePaymentMethodID"
        },
        "userID": {
          "$ref": "#/definitions/UserID"
        }
      },
      "x-go-package": "pr-reviewer/internal/model"
    },
    "UserID": {
      "description": "UserID is identifier for User",
      "type": "string",
      "x-go-package": "pr-reviewer/internal/model"
    },
...
    "genericError": {
      "description": "GenericError - represent error structure for generic error (we need this to make all our error response the same)",
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "x-go-name": "ErrorCode"
        },
        "data": {
          "x-go-name": "Data"
        },
        "message": {
          "type": "string",
          "x-go-name": "ErrorMessage"
        }
      },
      "x-go-name": "GenericError",
      "x-go-package": "pr-reviewer/internal/api/middlewares"
    }
  }
}